requirements.txt
diffusers
accelerate
torch
datasets
Pillow
requests
matplotlib
psutil
huggingface-hub